chore(lint): one lint/format gate on both trees (SPE-300)
Backend: ruff now selects E,W,F,I,UP,B,SIM,RUF,PT,ARG,ERA,T20,PTH,ASYNC,PERF with
per-file ignores for tests and alembic, and the gate (just lint, CI) covers src, tests
and alembic. One mechanical pass: imports sorted, Optional/Union -> PEP 604, datetime.UTC,
raise ... from, pathlib, list comprehensions; framework-required unused args renamed
with a leading underscore. ty pinned to 3.11 with possibly-missing-attribute promoted
to error. pytest filterwarnings = error (Starlette 422 rename ignored). Coverage gate
raised from the never-run 80 to the measured 90. New just deadcode recipe (vulture,
advisory).

Frontend: Prettier (no semi, single quotes, 100 cols, tailwind class order) formats the
tree; ESLint moves to recommendedTypeChecked with import-x (order, no-cycle), vitest and
testing-library rules for tests, eslint-config-prettier last. npm run lint = eslint +
knip; format:check and typecheck scripts; @vitest/coverage-v8 with thresholds at the
measured values (lines 72, statements 71, functions 66, branches 64). Floating promises
and unsafe-any sites in the API client, stream client and navigation handlers fixed;
Vite env typed in vite-env.d.ts.

Hooks: pre-commit runs ruff on backend src/tests/alembic and lint-staged (eslint --fix +
prettier) on staged frontend files; .secrets.baseline refreshed.

+++ backend/src/schemas/scoring_state.py
@@ -132,7 +132,7 @@ class DimensionScore(BaseModel):
             return 100 if self.level == 1 else 0
         if self.dimension == "demand":
             return DEMAND_LEVEL_TO_SCORE[self.level]
-        return int(round(self.expected / self.max_level * 100))
+        return round(self.expected / self.max_level * 100)
 
     def band(self) -> Band:
         return score_to_band(self.derived_score())
